+++ resolvers/webpack/test/config.js
@@ -36,4 +36,23 @@ describe("config", function () {
       .to.have.property('path')
       .and.equal(path.join(__dirname, 'config-extensions', 'src', 'main-module.js'))
   })
+
+  it("finds the first config with a resolve section", function () {
+    var settings = {
+      config: path.join(__dirname, './files/webpack.config.multiple.js'),
+    }
+
+    expect(resolve('main-module', file, settings)).to.have.property('path')
+        .and.equal(path.join(__dirname, 'files', 'src', 'main-module.js'))
+  })
+
+  it("finds the config at option config-index", function () {
+    var settings = {
+      config: path.join(__dirname, './files/webpack.config.multiple.js'),
+      'config-index': 2,
+    }
+
+    expect(resolve('foo', file, settings)).to.have.property('path')
+        .and.equal(path.join(__dirname, 'files', 'some', 'goofy', 'path', 'foo.js'))
+  })
 })