+++ resolvers/webpack/test/files/webpack.config.multiple.js
@@ -1,0 +1,20 @@
+var path = require('path')
+
+module.exports = [{
+  name: 'one',
+}, {
+  name: 'two',
+  resolve: {
+    root: path.join(__dirname, 'src'),
+    fallback: path.join(__dirname, 'fallback'),
+  },
+}, {
+  name: 'three',
+  resolve: {
+    alias: {
+      'foo': path.join(__dirname, 'some', 'goofy', 'path', 'foo.js'),
+    },
+    modulesDirectories: ['node_modules', 'bower_components'],
+    root: path.join(__dirname, 'src'),
+  },
+}]